+++ engine/runtime/atlas/Cargo.toml
@@ -14,6 +14,7 @@ version.workspace = true
 image.workspace = true
 rstar.workspace = true
 thiserror.workspace = true
+tracing.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
