+++ engine/runtime/atlas/src/lib.rs
@@ -1,6 +1,7 @@
+mod blit;
 mod damage;
 mod error;
-mod pack;
+mod plan;
 mod skyline;
 
 use std::collections::HashMap;
@@ -8,6 +9,7 @@ use std::path::PathBuf;
 
 use damage::{collect_damage, TextureDamage};
 pub use error::{AtlasError, Result};
+pub use plan::plan_layout;
 use image::RgbaImage;
 
 pub type PolygonUVs = Vec<[f64; 2]>;
@@ -122,7 +124,7 @@ fn remap_polygon_uvs(
 fn build_remapped_uvs(
     materials: &[TextureMaterial],
     damage_list: &[(PathBuf, TextureDamage)],
-    texture_frames: &pack::TextureFrames,
+    texture_frames: &blit::TextureFrames,
     downsample: u32,
     atlas_size: (f64, f64),
 ) -> Vec<Option<TextureUVs>> {
@@ -161,57 +163,6 @@ fn build_remapped_uvs(
         .collect()
 }
 
-/// Returns the minimum k such that `side / 2^k <= max`, or `MAX_DOWNSAMPLE_K + 1` if none exists.
-fn needed_k(side: u32, max: u32) -> u32 {
-    if side <= max {
-        return 0;
-    }
-    let ratio = side.div_ceil(max);
-    if ratio > (1u32 << MAX_DOWNSAMPLE_K) {
-        return MAX_DOWNSAMPLE_K + 1;
-    }
-    ratio.next_power_of_two().trailing_zeros()
-}
-
-/// Compute the layout for a set of textures given only their dimensions.
-/// No image files are read; no pixels are blitted.
-/// Useful for efficiency benchmarks and layout-only unit tests.
-pub fn plan_layout(dims: &[(u32, u32)], max_atlas_size: u32) -> Result<LayoutPlan> {
-    if dims.is_empty() {
-        return Ok(LayoutPlan {
-            atlas_width: 1,
-            atlas_height: 1,
-            downsample: 1,
-            placements: vec![],
-        });
-    }
-    // `virtual_w/h` is the estimated canvas in original (pre-downsampling) pixel space.
-    // Doubling grows the canvas at k=0; once the larger dimension exceeds max_atlas_size,
-    // needed_k increments to keep the physical atlas within bounds. Both k and canvas are derived.
-    let (mut virtual_w, mut virtual_h) = pack::estimate_atlas_size_from_dims(dims, 0);
-    loop {
-        let k = needed_k(virtual_w.max(virtual_h), max_atlas_size);
-        if k > MAX_DOWNSAMPLE_K {
-            break;
-        }
-        let canvas = (virtual_w.min(max_atlas_size), virtual_h.min(max_atlas_size));
-        if let Some((used_w, used_h, placements)) = pack::try_layout_rects(dims, k, canvas) {
-            return Ok(LayoutPlan {
-                atlas_width: used_w,
-                atlas_height: used_h,
-                downsample: 1u32 << k,
-                placements,
-            });
-        }
-        virtual_w = virtual_w.saturating_mul(2);
-        virtual_h = virtual_h.saturating_mul(2);
-    }
-    Err(AtlasError::builder(format!(
-        "Texture atlas does not fit within {}x{} even at downsample factor 2^{}",
-        max_atlas_size, max_atlas_size, MAX_DOWNSAMPLE_K
-    )))
-}
-
 /// Pack `materials` into an atlas image and return remapped UVs.
 /// `remapped_uvs[i]` is `Some(remapped_uvs)` if `materials[i]` was packed, `None` if excluded.
 pub fn build_atlas(materials: &[TextureMaterial], max_atlas_size: u32) -> Result<BuiltAtlas> {
@@ -229,7 +180,7 @@ pub fn build_atlas(materials: &[TextureMaterial], max_atlas_size: u32) -> Result
     let plan = plan_layout(&dims, max_atlas_size)?;
 
     // Stage 3: blit using pre-computed placements — no second layout pass.
-    let (atlas, texture_frames) = pack::blit(
+    let (atlas, texture_frames) = blit::blit(
         &damage_list,
         (plan.atlas_width, plan.atlas_height),
         &plan.placements,
